+++ src/matter/export.ts
@@ -6,8 +6,3 @@ export * from '@matter/main/behaviors';
 // export * from '@matter/main/clusters';
 export { ExposedFabricInformation, FabricAction, MdnsService, PaseClient } from '@matter/main/protocol';
 export { AttributeElement, ClusterElement, ClusterModel, CommandElement, EventElement, FieldElement } from '@matter/main/model';
-
-// Matterbridge
-export * from '../matterbridgeDeviceTypes.js';
-export * from '../matterbridgeEndpoint.js';
-export * from '../matterbridgeBehaviors.js';